Converted the real_time value field to float.  Django can not convert floats to decimal.

+++ depth/models.py
@@ -319,7 +319,7 @@ class ToolMWDRealTime(models.Model) :
     well = models.ForeignKey(Well)    
     time_stamp = models.DateTimeField( db_index=True)
     type = models.CharField(max_length=32, choices = VALUE_TYPE_CHOICES, db_index=True)
-    value = models.DecimalField(max_digits=10, decimal_places=3, null=True, blank=True)
+    value = models.FloatField(null=True, blank=True)
     value_x = models.IntegerField(blank=True, null=True)
     value_y = models.IntegerField(blank=True, null=True)
     value_z = models.IntegerField(blank=True, null=True)
